services/requirements.txt
fastapi
uvicorn
python-multipart
python-magic
markitdown
